Add character contract items & bids updater and fix character udpaters.
* Endpoints have since been paginated
* Contracts filters for items & bids are now filtered by
  character_id

+++ src/database/migrations/2018_01_07_134346_create_contract_details_table.php
@@ -42,7 +42,7 @@ class CreateContractDetailsTable extends Migration
             $table->double('reward')->nullable();
             $table->double('collateral')->nullable();
             $table->double('buyout')->nullable();
-            $table->float('volume')->nullable();
+            $table->double('volume')->nullable();
 
             $table->index('issuer_id');
             $table->index('issuer_corporation_id');
